requirements.txt
flask
flask-sqlalchemy
flask-cors
python-dotenv
PyJWT
psycopg2-binary
